Enable I2C
I do not like the way the block markers don't have a leading
blank line. I found a way to add one:
http://mrtango.planetcrazy.de/ansible-add-an-empty-line-before-a-blockinfile.html

I improved the method slightly (more DRY). It's still fragile and will
break if the marker format is changed. It can be improved, but not
solved, with regular expresions.

It's acceptable. The correct fix is not worry about the line break, and
the reduced readability that comes with.

Trying to "trick" ansible into doing "the right thing" by prepending a
newline to the marker merely results in the block getting replicated.
Nice try, though.