Support arbitrary local user
I was working on a Pi4 as the controller. Ansible would take over ten
seconds to simply respond to `ansible --version`. I am now using a
Debian 10 WSL as the ansible controller, and the username is no longer
"pi". So if you aren't pi this will work for you, too.

The localhost.yml playbook still works. At least for me.