Add brick mount points
```
volume create: cifs: failed: The brick node1:/data/glusterfs/cifs/brick1
is a mount point. Please create a sub-directory under the mount point
and use that as the brick directory. Or use 'force' at the end of the
command if you want to override this behavior.
```

We don't want to _force_ it, do we?[1]

[1]: No. We don't.