Fix gluster idempotency
If a subset of the cluster was run through site.yml, it would fail due
to redundancy requirements, even if the cluster exists.

Refactored to query if the requested cluster exists, and skip it.

Reboot refactored to mark OS as reboot-required, and reboot at end of
site.yml; two changes require reboots and rebooting twice is silly.